+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.1] - 2023-01-24
+### Added
+- More continuous haptic envelope generation parameters
+
+### Changed
+- Ensured that FFT chunk size parameter for continuous envelope generation will always be a power of 2
+
+### Fixed
+- Audio waveform texture timing was corrected
+
+
 ## [0.5.0] - 2023-01-22
 ### Added
 - Basic algorithms to automatically generate haptics: